requirements.txt
beautifulsoup4
requests
pushbullet.py

